Добавила hover на конопку Выйти
