+++ src/components/ExitModal.vue
@@ -1,35 +1,42 @@
 <template>
-  <div>
-    <div class="pop-exit" id="popExit">
-      <div class="pop-exit__container">
-        <div class="pop-exit__block">
-          <div class="pop-exit__ttl">
-            <h2>Выйти из аккаунта?</h2>
-          </div>
-          <form class="pop-exit__form" id="formExit" action="#">
-            <div class="pop-exit__form-group">
-              <BaseButton class="pop-exit__exit-yes _hover01" id="exitYes">
-                <a href="modal/signin.html">Да, выйти</a>
-              </BaseButton>
-              <button class="pop-exit__exit-no _hover03" id="exitNo">
-                <a href="main.html">Нет, остаться</a>
-              </button>
-            </div>
-          </form>
+  <div class="pop-exit" id="popExit">
+    <div class="pop-exit__container">
+      <div class="pop-exit__block">
+        <div class="pop-exit__ttl">
+          <h2>Выйти из аккаунта?</h2>
         </div>
+        <form class="pop-exit__form" id="formExit" action="#">
+          <div class="pop-exit__form-group">
+            <RouterLink to="/sign-in">
+              <BaseButton @click="logout" class="pop-exit__exit-yes _hover01">Да, выйти</BaseButton>
+            </RouterLink>
+
+            <RouterLink to="/">
+              <BaseButton class="pop-exit__exit-no _hover03">Нет, остаться</BaseButton>
+            </RouterLink>
+          </div>
+        </form>
       </div>
     </div>
   </div>
 </template>
 
 <script setup>
-import BaseButton from './BaseButton.vue';
+import { RouterLink, useRouter } from 'vue-router'
+import BaseButton from './BaseButton.vue'
 
+const router = useRouter()
+
+function logout(e) {
+  e.preventDefault()
+  localStorage.removeItem('userInfo')
+  router.push('/sign-in')
+}
 </script>
 
 <style scoped>
 .pop-exit {
-  display: none;
+  display: block;
   width: 100%;
   height: 100%;
   min-width: 320px;
@@ -108,7 +115,7 @@ import BaseButton from './BaseButton.vue';
   line-height: 21px;
   font-weight: 500;
   letter-spacing: -0.14px;
-  color: #ffffff;
+  color: #565eef;
 }
 .pop-exit__exit-no a {
   width: 100%;
@@ -122,9 +129,9 @@ import BaseButton from './BaseButton.vue';
   width: 100%;
   display: flex;
   align-items: center;
-  justify-content: space-between;
+  justify-content: center;
 }
-.pop-exit:target{
+.pop-exit:target {
   display: block;
 }
 @media only screen and (max-width: 375px) {
@@ -144,6 +151,5 @@ import BaseButton from './BaseButton.vue';
   .pop-exit__form-group {
     display: block;
   }
-} 
+}
 </style>
-
